refactor(desktop/timeline): move validateGolarian to calendar, rename session-persistence
- Remove validateGolarian from session-domain.ts; use tryParseDate from
  calendar/golarian instead (it already existed and does the same thing).
  validateSessionBuffer now calls tryParseDate directly.
- Rename interactions/session-persistence.ts → session-transforms.ts since
  the file contains pure array transforms, not I/O. Update all imports and
  test file accordingly.
