+++ desktop/src/renderer/timeline/interactions/session-transforms.ts
